+++ CHANGELOG.md
@@ -31,6 +31,8 @@ v1.0.0 (in development)
 - Renamed `DistributionPackage.from_pep691_details()` to `from_json_data()`
 - `PyPISimple.stream_project_names()` now accepts JSON responses
 - Use pydantic internally to parse JSON responses
+- Added constants for passing to `PyPISimple` and its methods in order to
+  specify the `Accept` header to send
 
 v0.10.0 (2022-06-30)
 --------------------