+++ docs/changelog.rst
@@ -43,6 +43,8 @@ v1.0.0 (in development)
   `~DistributionPackage.from_json_data()`
 - `PyPISimple.stream_project_names()` now accepts JSON responses
 - Use pydantic internally to parse JSON responses
+- Added constants for passing to `PyPISimple` and its methods in order to
+  specify the :mailheader:`Accept` header to send
 
 v0.10.0 (2022-06-30)
 --------------------